changed cost field to DecimalField

+++ booking/models.py
@@ -12,7 +12,7 @@ class Event(models.Model):
     max_participants = models.PositiveIntegerField(null=True, blank=True, help_text="Leave blank if no max number of participants")
     contact_person = models.CharField(max_length=255, default="Gwen Burns")
     contact_email = models.EmailField(default="thewatermelonstudio@hotmail.com")
-    cost = models.PositiveIntegerField(default=0)
+    cost = models.DecimalField(default=0, max_digits=8, decimal_places=2)
     payment_open = models.BooleanField(default=False)
     payment_info = models.TextField(blank=True)
     payment_due_date = models.DateTimeField(null=True, blank=True)
